serial: move `RingBuffer` to its own module
